Split sorting algorithms into individual files
Each algorithm (bubble, insertion, binaryInsertion, selection, merge,
mergeAdaptive, quick) now lives in its own file under algorithms/,
containing its generator function, helpers, and metadata. An index.js
re-exports them all; app.jsx imports from there and contains only
rendering logic.
