+++ src/pages/tools/sorting-comparator/_app/algorithms/index.js
@@ -1,0 +1,19 @@
+import bubble from './bubble.js';
+import insertion from './insertion.js';
+import binaryInsertion from './binaryInsertion.js';
+import selection from './selection.js';
+import merge from './merge.js';
+import mergeAdaptive from './mergeAdaptive.js';
+import quick from './quick.js';
+
+const ALGORITHMS = {
+  bubble,
+  insertion,
+  binaryInsertion,
+  selection,
+  merge,
+  mergeAdaptive,
+  quick,
+};
+
+export default ALGORITHMS;
