+++ src/pages/tools/sorting-comparator/_app/app.jsx
@@ -1,245 +1,7 @@
 import React, { useState, useRef } from 'react';
 import { createRoot } from 'react-dom/client';
 import './app.css';
-
-// ——— Sorting Generators ———
-// Protocol: yield { a, b } → receive cmp (number):
-//   cmp > 0  →  a is greater (a should come later in ascending order)
-//   cmp = 0  →  equal
-//   cmp < 0  →  a is smaller
-// Returns the sorted array in ascending order.
-
-function* bubbleSort(items) {
-  const arr = [...items];
-  for (let i = 0; i < arr.length - 1; i++) {
-    let swapped = false;
-    for (let j = 0; j < arr.length - i - 1; j++) {
-      const cmp = yield { a: arr[j], b: arr[j + 1] };
-      if (cmp > 0) {
-        [arr[j], arr[j + 1]] = [arr[j + 1], arr[j]];
-        swapped = true;
-      }
-    }
-    if (!swapped) break; // Early exit: no swaps means sorted
-  }
-  return arr;
-}
-
-function* insertionSort(items) {
-  const arr = [...items];
-  for (let i = 1; i < arr.length; i++) {
-    let j = i;
-    while (j > 0) {
-      const cmp = yield { a: arr[j - 1], b: arr[j] };
-      if (cmp > 0) {
-        [arr[j - 1], arr[j]] = [arr[j], arr[j - 1]];
-        j--;
-      } else {
-        break; // Covers both equal (cmp=0) and less-than (cmp<0)
-      }
-    }
-  }
-  return arr;
-}
-
-function* selectionSort(items) {
-  const arr = [...items];
-  for (let i = 0; i < arr.length - 1; i++) {
-    let minIdx = i;
-    for (let j = i + 1; j < arr.length; j++) {
-      const cmp = yield { a: arr[minIdx], b: arr[j] };
-      if (cmp > 0) minIdx = j; // arr[minIdx] > arr[j] → new minimum found
-    }
-    if (minIdx !== i) [arr[i], arr[minIdx]] = [arr[minIdx], arr[i]];
-  }
-  return arr;
-}
-
-function* mergeSort(items) {
-  const arr = [...items];
-  const n = arr.length;
-  for (let width = 1; width < n; width *= 2) {
-    for (let lo = 0; lo < n; lo += 2 * width) {
-      const mid = Math.min(lo + width, n);
-      const hi = Math.min(lo + 2 * width, n);
-      const left = arr.slice(lo, mid);
-      const right = arr.slice(mid, hi);
-      if (right.length === 0) continue;
-      let li = 0, ri = 0;
-      const merged = [];
-      while (li < left.length && ri < right.length) {
-        const cmp = yield { a: left[li], b: right[ri] };
-        if (cmp > 0) {
-          merged.push(right[ri++]); // left > right → take right (smaller)
-        } else {
-          merged.push(left[li++]); // cmp <= 0 → take left (stable: left wins on equal)
-        }
-      }
-      while (li < left.length) merged.push(left[li++]);
-      while (ri < right.length) merged.push(right[ri++]);
-      for (let i = 0; i < merged.length; i++) arr[lo + i] = merged[i];
-    }
-  }
-  return arr;
-}
-
-// Binary insertion sort: uses binary search to find the insertion position.
-// Each element is compared via binary search in O(log n) comparisons.
-// The same value pair is never compared twice — every comparison yields new info.
-function* binaryInsertionSort(items) {
-  const arr = [...items];
-  for (let i = 1; i < arr.length; i++) {
-    let lo = 0, hi = i;
-    while (lo < hi) {
-      const mid = Math.floor((lo + hi) / 2);
-      const cmp = yield { a: arr[i], b: arr[mid] };
-      // cmp < 0: arr[i] < arr[mid] → search left half
-      // cmp >= 0: arr[i] >= arr[mid] → search right half (stable: equal goes right)
-      if (cmp < 0) hi = mid; else lo = mid + 1;
-    }
-    // Shift arr[lo..i-1] right and insert arr[i] at position lo
-    const val = arr[i];
-    for (let j = i; j > lo; j--) arr[j] = arr[j - 1];
-    arr[lo] = val;
-  }
-  return arr;
-}
-
-// Adaptive merge sort: checks if left[last] ≤ right[first] before each merge.
-// If the check passes, the entire merge is skipped with just 1 comparison.
-// On nearly-sorted or already-sorted data this dramatically reduces comparisons.
-// When a full merge is needed, the pre-check result is reused — never double-asked.
-function* mergeSortAdaptive(items) {
-  const arr = [...items];
-  const n = arr.length;
-  for (let width = 1; width < n; width *= 2) {
-    for (let lo = 0; lo < n; lo += 2 * width) {
-      const mid = Math.min(lo + width, n);
-      const hi = Math.min(lo + 2 * width, n);
-      const left = arr.slice(lo, mid);
-      const right = arr.slice(mid, hi);
-      if (right.length === 0) continue;
-      // Pre-sorted check: 1 comparison can skip the entire merge
-      const preCmp = yield { a: left[left.length - 1], b: right[0] };
-      if (preCmp <= 0) continue; // left[last] ≤ right[first] → already in order
-      // Full merge needed; reuse preCmp when we reach (li=last, ri=0) in the loop
-      let li = 0, ri = 0;
-      const merged = [];
-      while (li < left.length && ri < right.length) {
-        let cmp;
-        if (li === left.length - 1 && ri === 0) {
-          cmp = preCmp; // Already know: left[last] > right[0]
-        } else {
-          cmp = yield { a: left[li], b: right[ri] };
-        }
-        if (cmp > 0) {
-          merged.push(right[ri++]);
-        } else {
-          merged.push(left[li++]);
-        }
-      }
-      while (li < left.length) merged.push(left[li++]);
-      while (ri < right.length) merged.push(right[ri++]);
-      for (let i = 0; i < merged.length; i++) arr[lo + i] = merged[i];
-    }
-  }
-  return arr;
-}
-
-// 3-way quicksort (Dutch National Flag partition).
-// Partitions into: [< pivot] [== pivot] [> pivot]
-// All equal-to-pivot elements are placed in O(n) and need no further sorting.
-function* quickSort(items) {
-  const arr = [...items];
-  function* qs(lo, hi) {
-    if (hi - lo <= 1) return;
-    // Random pivot → move to front
-    const pivotIdx = lo + Math.floor(Math.random() * (hi - lo));
-    [arr[lo], arr[pivotIdx]] = [arr[pivotIdx], arr[lo]];
-    const pivotVal = arr[lo];
-    // Invariant:
-    //   arr[lo..lt-1]  < pivot  (less-than zone)
-    //   arr[lt..i-1]  == pivot  (equal zone, starts with just the pivot at lt=lo)
-    //   arr[i..gt-1]  = unclassified
-    //   arr[gt..hi-1] > pivot  (greater-than zone)
-    let lt = lo, gt = hi, i = lo + 1;
-    while (i < gt) {
-      const cmp = yield { a: arr[i], b: pivotVal };
-      if (cmp < 0) {
-        [arr[i], arr[lt]] = [arr[lt], arr[i]];
-        lt++; i++;
-      } else if (cmp > 0) {
-        gt--;
-        [arr[i], arr[gt]] = [arr[gt], arr[i]];
-        // Don't advance i — the swapped-in element needs classification
-      } else {
-        i++; // Equal to pivot: absorb into the equal zone
-      }
-    }
-    yield* qs(lo, lt);  // Sort the less-than partition
-    yield* qs(gt, hi);  // Sort the greater-than partition
-    // [lt, gt) are all == pivot — already in final position, no recursion needed
-  }
-  yield* qs(0, arr.length);
-  return arr;
-}
-
-// ——— Helpers ———
-
-// Best-case comparisons for binary insertion sort (sorted input: always go right)
-// Each position i costs floor(log2(i+1)) comparisons.
-function binaryInsertBest(n) {
-  let count = 0;
-  for (let i = 1; i < n; i++) count += Math.floor(Math.log2(i + 1));
-  return count;
-}
-
-// Worst-case comparisons for binary insertion sort (reverse-sorted: always go left)
-// Each position i costs ceil(log2(i+1)) comparisons.
-function binaryInsertWorst(n) {
-  let count = 0;
-  for (let i = 1; i < n; i++) count += Math.ceil(Math.log2(i + 1));
-  return count;
-}
-
-// Number of actual merges in bottom-up merge sort (right half non-empty)
-function countMerges(n) {
-  let count = 0;
-  for (let w = 1; w < n; w *= 2)
-    for (let lo = 0; lo < n; lo += 2 * w) {
-      const mid = Math.min(lo + w, n);
-      const hi = Math.min(lo + 2 * w, n);
-      if (hi > mid) count++;
-    }
-  return count;
-}
-
-// Maximum comparisons for bottom-up merge sort on n elements
-function mergeUpperBound(n) {
-  let count = 0;
-  for (let w = 1; w < n; w *= 2)
-    for (let lo = 0; lo < n; lo += 2 * w) {
-      const mid = Math.min(lo + w, n);
-      const hi = Math.min(lo + 2 * w, n);
-      const r = hi - mid;
-      if (r > 0) count += (mid - lo) + r - 1;
-    }
-  return count;
-}
-
-// Minimum comparisons for bottom-up merge sort on n elements
-// (when each merge exhausts the shorter half first)
-function mergeBestCase(n) {
-  let count = 0;
-  for (let w = 1; w < n; w *= 2)
-    for (let lo = 0; lo < n; lo += 2 * w) {
-      const mid = Math.min(lo + w, n);
-      const hi = Math.min(lo + 2 * w, n);
-      const l = mid - lo, r = hi - mid;
-      if (r > 0) count += Math.min(l, r);
-    }
-  return count;
-}
+import ALGORITHMS from './algorithms/index.js';
 
 function shuffle(arr) {
   const a = [...arr];
@@ -257,185 +19,6 @@ function startSort(algoKey, items) {
     return { gen, done: true, sorted: first.value, comparison: null, completed: 0 };
   return { gen, done: false, sorted: null, comparison: first.value, completed: 0 };
 }
-
-// ——— Algorithm metadata ———
-
-const ALGORITHMS = {
-  bubble: {
-    name: 'Bubble Sort',
-    fn: bubbleSort,
-    complexity: 'O(n²)',
-    getTotal: n => n * (n - 1) / 2,
-    exact: false,
-    stable: true,
-    noRepeatPairs: false,
-    description:
-      'Repeatedly scans adjacent pairs and swaps them if out of order. Early exit makes it O(n) on sorted data. Downside for humans: the same adjacent pairs re-appear in every pass, so you\'ll see familiar faces over and over.',
-    estimates: {
-      best:  n => Math.max(0, n - 1),
-      avg:   n => Math.round(n * (n - 1) / 4),
-      worst: n => n * (n - 1) / 2,
-      bestLabel:  'already sorted',
-      worstLabel: 'reverse sorted',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'great', note: 'Early exit kicks in — just a few passes' },
-      { label: 'Random data',     rating: 'poor',  note: 'Full n² comparisons in practice' },
-      { label: 'Reverse sorted',  rating: 'poor',  note: 'Maximum work — every pair must bubble up' },
-      { label: 'Many duplicates', rating: 'ok',    note: 'Stable — equal elements never swap' },
-    ],
-  },
-
-  insertion: {
-    name: 'Insertion Sort',
-    fn: insertionSort,
-    complexity: 'O(n²)',
-    getTotal: n => n * (n - 1) / 2,
-    exact: false,
-    stable: true,
-    noRepeatPairs: false,
-    description:
-      'Builds a sorted prefix one element at a time, shifting each new item left until it finds its spot. Best O(n²) algorithm for nearly-sorted data, but linear search means the same pair of values can appear again with duplicate items.',
-    estimates: {
-      best:  n => Math.max(0, n - 1),
-      avg:   n => Math.round(n * (n - 1) / 4),
-      worst: n => n * (n - 1) / 2,
-      bestLabel:  'already sorted',
-      worstLabel: 'reverse sorted',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'great', note: 'Only 1–2 comparisons per element' },
-      { label: 'Random data',     rating: 'poor',  note: '~n²/4 comparisons on average' },
-      { label: 'Reverse sorted',  rating: 'poor',  note: 'Each element must shift all the way left' },
-      { label: 'Many duplicates', rating: 'ok',    note: 'Stable; but same value pair may reappear' },
-    ],
-  },
-
-  binaryInsertion: {
-    name: 'Binary Insertion Sort',
-    fn: binaryInsertionSort,
-    complexity: 'O(n log n)',
-    getTotal: n => binaryInsertWorst(n),
-    exact: false,
-    stable: true,
-    noRepeatPairs: true,
-    description:
-      'Like insertion sort, but uses binary search to locate the insertion position. Every comparison is unique — you will never see the same pair twice. O(log n) comparisons per element makes for a very fair, varied experience.',
-    estimates: {
-      best:  n => binaryInsertBest(n),
-      avg:   n => Math.round((binaryInsertBest(n) + binaryInsertWorst(n)) / 2),
-      worst: n => binaryInsertWorst(n),
-      bestLabel:  'already sorted',
-      worstLabel: 'reverse sorted',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'ok',    note: 'Binary search does the same work regardless of order' },
-      { label: 'Random data',     rating: 'great', note: 'No repeated pairs — each comparison tells you something new' },
-      { label: 'Reverse sorted',  rating: 'ok',    note: 'Same binary-search count as sorted — no worst case' },
-      { label: 'Many duplicates', rating: 'great', note: 'Stable and handles equals gracefully' },
-    ],
-  },
-
-  selection: {
-    name: 'Selection Sort',
-    fn: selectionSort,
-    complexity: 'O(n²)',
-    getTotal: n => n * (n - 1) / 2,
-    exact: true, // Always exactly n(n-1)/2 comparisons
-    stable: false,
-    noRepeatPairs: false,
-    description:
-      'Finds the minimum of the unsorted portion and moves it into place, every round. Makes exactly n(n−1)/2 comparisons regardless of input. Tedious for humans: the current minimum dominates every comparison in a round.',
-    estimates: {
-      best:  n => n * (n - 1) / 2,
-      avg:   n => n * (n - 1) / 2,
-      worst: n => n * (n - 1) / 2,
-      bestLabel:  'always the same',
-      worstLabel: 'always the same',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'poor', note: 'No benefit — always full n² comparisons' },
-      { label: 'Random data',     rating: 'ok',   note: 'Predictable but quadratic' },
-      { label: 'Reverse sorted',  rating: 'ok',   note: 'Same work as any other input' },
-      { label: 'Many duplicates', rating: 'ok',   note: 'Not stable, but comparison count unchanged' },
-    ],
-  },
-
-  merge: {
-    name: 'Merge Sort',
-    fn: mergeSort,
-    complexity: 'O(n log n)',
-    getTotal: n => mergeUpperBound(n),
-    exact: false,
-    stable: true,
-    noRepeatPairs: true,
-    description:
-      'Splits in half, sorts each half, merges the results. Guaranteed O(n log n) in all cases. Each pair of values is compared at most once, so you\'ll never see a repeat. For sorted inputs, try Adaptive Merge Sort to skip unnecessary merges.',
-    estimates: {
-      best:  n => mergeBestCase(n),
-      avg:   n => Math.max(0, Math.round(n * Math.log2(Math.max(n, 1)) - n + 1)),
-      worst: n => mergeUpperBound(n),
-      bestLabel:  'one side exhausted first in each merge',
-      worstLabel: 'each merge compares every element',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'ok',    note: 'Still performs all merges — no shortcut' },
-      { label: 'Random data',     rating: 'great', note: 'Optimal O(n log n) guaranteed' },
-      { label: 'Reverse sorted',  rating: 'great', note: 'Same as random — no worst case' },
-      { label: 'Many duplicates', rating: 'great', note: 'Stable — equal elements maintain original order' },
-    ],
-  },
-
-  mergeAdaptive: {
-    name: 'Adaptive Merge Sort',
-    fn: mergeSortAdaptive,
-    complexity: 'O(n log n)',
-    getTotal: n => mergeUpperBound(n),
-    exact: false,
-    stable: true,
-    noRepeatPairs: true,
-    description:
-      'Merge sort with a pre-sorted check before each merge: "is left[last] ≤ right[first]?" One comparison can skip an entire merge group. On already-sorted data it uses only one comparison per merge — ideal when you suspect your list is nearly in order.',
-    estimates: {
-      best:  n => countMerges(n),
-      avg:   n => Math.round((countMerges(n) + mergeUpperBound(n)) / 2),
-      worst: n => mergeUpperBound(n),
-      bestLabel:  'already sorted — each merge skipped in 1 comparison',
-      worstLabel: 'each merge needs full comparison work',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'great', note: 'Pre-check skips most or all merges in 1 comparison each' },
-      { label: 'Random data',     rating: 'great', note: 'O(n log n) guaranteed, never asks same pair twice' },
-      { label: 'Reverse sorted',  rating: 'ok',    note: 'Full merges needed, but still no repeated pairs' },
-      { label: 'Many duplicates', rating: 'great', note: 'Stable — equal elements maintain original order' },
-    ],
-  },
-
-  quick: {
-    name: 'Quicksort (3-way)',
-    fn: quickSort,
-    complexity: 'O(n log n) avg',
-    getTotal: n => n * (n - 1) / 2,
-    exact: false,
-    stable: false,
-    noRepeatPairs: false,
-    description:
-      'Picks a random pivot and partitions into smaller, equal, and larger groups. Equal elements are placed in one pass and need no further sorting. Downside for humans: the pivot dominates every comparison within a partition round.',
-    estimates: {
-      best:  n => Math.max(0, Math.round(n * Math.log2(Math.max(n, 1)))),
-      avg:   n => Math.max(0, Math.round(1.39 * n * Math.log2(Math.max(n, 1)))),
-      worst: n => n * (n - 1) / 2,
-      bestLabel:  'perfectly balanced partitions',
-      worstLabel: 'always picks worst pivot',
-    },
-    scenarios: [
-      { label: 'Nearly sorted',   rating: 'ok',    note: 'Random pivot prevents sorted-input worst case' },
-      { label: 'Random data',     rating: 'great', note: 'Best practical performance in most benchmarks' },
-      { label: 'Reverse sorted',  rating: 'ok',    note: 'Random pivot avoids O(n²) degenerate behavior' },
-      { label: 'Many duplicates', rating: 'great', note: '3-way partition: all duplicates placed in O(n)' },
-    ],
-  },
-};
 
 // ——— AlgoInfo component ———
